+++ src/index.ts
@@ -1,6 +1,6 @@
 import app from './app';
 import * as dotenv from 'dotenv';
-import logger from './modules/logging/winston';
+import logger from './utils/logging/winston';
 import {Prisma }from './utils/db';
 import { Application } from 'express';
